Allow communication in push mode
By using the command `"Sn\n"` (where `n` is an integer specifying the
timeout in milliseconds), the protocol will be switched in push mode. In
this mode, the status of each encoder is output whenever it changes. The
maximum time between two refreshes is `n` milliseconds.

By default the protocol starts in slave mode, i.e. the same mode
activated by an `"S0\n"` command. In this mode the Arduino program will
respond only to specific requests.