Remove many `static`s
ardecoder.ino is not proper C language... it is a bastard C++ variant
crafted for non-programmers. As such I removed many static keywords that
can potentially get in the way of the arduino tools.

When I have more experience I am planning to rewrite this crap in proper
good ol' C.